:goal_net: re-name error file and re-factor the exception class
