+++ helpers/pyband/pyband/exceptions.py
@@ -1,0 +1,50 @@
+class EmptyRequestMsgError(Exception):
+    pass
+
+
+class QueryError(Exception):
+    pass
+
+
+class NegativeIntegerError(Exception):
+    pass
+
+
+class ValueTooLargeError(Exception):
+    pass
+
+
+class InsufficientCoinError(Exception):
+    pass
+
+
+class EmptyMsgError(Exception):
+    pass
+
+
+class NotFoundError(Exception):
+    pass
+
+
+class UndefinedError(Exception):
+    pass
+
+
+class DecodeError(Exception):
+    pass
+
+
+class ConvertError(Exception):
+    pass
+
+
+class UnsuccessfulCallError(Exception):
+    pass
+
+
+class CreateError(Exception):
+    pass
+
+
+class SchemaError(Exception):
+    pass
