+++ helpers/pyband/pyband/wallet.py
@@ -6,7 +6,7 @@ from bip32 import BIP32
 from ecdsa import SigningKey, VerifyingKey, SECP256k1, BadSignatureError
 from ecdsa.util import sigencode_string_canonize
 from mnemonic import Mnemonic
-from .error import ConvertError, DecodeError
+from .exceptions import ConvertError, DecodeError
 
 BECH32_PUBKEY_ACC_PREFIX = "bandpub"
 BECH32_PUBKEY_VAL_PREFIX = "bandvaloperpub"
